Add `Meta.ParseError` detail field
Here we add a `detail` field to `Meta.ParseError`, but retain the
`msg::String` field for compatibility. `showerror(::ParseError)` defers
to the `detail` field if it's present.

This allows us to still throw `Meta.ParseError` from `Meta.parse` for
compatibility, but allow more expressivity when `detail` is set to an
exception type like `JuliaSyntax.ParseError`.